move page out of component directory
