+++ src/shared/app.jsx
@@ -7,10 +7,10 @@ import Helmet from 'react-helmet';
 
 import { APP_NAME } from './config';
 import Nav from './component/nav';
-import HomePage from './component/page/home';
-import HelloPage from './component/page/hello';
-import HelloAsyncPage from './component/page/hello-async';
-import NotFoundPage from './component/page/not-found';
+import HomePage from './page/home';
+import HelloPage from './page/hello';
+import HelloAsyncPage from './page/hello-async';
+import NotFoundPage from './page/not-found';
 import {
   HOME_PAGE_ROUTE,
   HELLO_PAGE_ROUTE,
