+++ src/shared/page/home/index.js
